+++ tests/python/test_moe.py
@@ -178,7 +178,8 @@ def test_llama4_moe_thunderfx():
     # as it doesn't have a grad-rule for the same.
     model.requires_grad_(False)
 
-    batch_size, seq_len = 1, 2048
+    #batch_size, seq_len = 1, 2048
+    batch_size, seq_len = 1, 1024
     inp = torch.randn(
         batch_size, seq_len, config.hidden_size, dtype=torch.bfloat16, device="cuda"
     )
